requirements.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf_yasg==1.21.8
gunicorn==21.2.0
psycopg2-binary==2.9.9
postmarker==1.0
whitenoise==6.6.0
django-cors-headers==4.3.1
django-filter==23.5
Pillow==10.2.0
dj-database-url==2.1.0
